[sgen] Make binary protocol thread-safe for real.
There was still an issue with thread-safety of binary protocols: If
the world was stopped while a thread was in the process of using a
binary protocol buffer, the GC would write all buffers to the file and
release the memory, making the thread crash once it restarted.

We fix this here by keeping a count of how many threads currently use
the buffers.  If they are in use when we try to save and free them, we
don't.  They'll be saved and freed eventually.


Commit migrated from https://github.com/mono/mono/commit/6d5c3cbc132f4b3166d8641216d9744d6035d045