mdoc 2.0-only profile build support, take 2!
Further discussion suggested that, even if mdoc 2.0 works, it hasn't gotten
widespread testing (as it's been a 4.0 app for months), so it would be
preferable to keep it as a 4.0 app.

However, it still needs to be built under 2.0 so that a --with-profile4=no
build can...build...without errors.

Thus, the compromise: build mdoc twice, once for each profile.  This will
ensure that *some* version of mdoc is around when mcs/docs is built.


Commit migrated from https://github.com/mono/mono/commit/313633fc67cf1479514019a56143e5ad1643d28f