[sgen] Only load an object's vtable word if absolutely necessary in M&S.
Until now we always loaded the vtable word in the mark/copy function.
It is not necessary in most cases for the fixed heap collector,
however.  If the object is on the small object major heap and its
block size is not evacuated we don't need to load it.


Commit migrated from https://github.com/mono/mono/commit/77c53b9b1857caa83351913d7eff05091054974b