[sgen] Two carefully placed prefetches.
Prefetch an object's vtable word when it's put on the gray stack,
because it's likely to be scanned soon.

Also prefetch it right before calling the mark/copy function in the
dynamic heap M&S to win a few cycles on the load.  Don't do it for the
fixed heap because there it's unlikely we'll need the vtable anyway,
so the cache can be put to good use elsewhere.


Commit migrated from https://github.com/mono/mono/commit/b8f9fee496ef626533b1fc9d88f995cfeb9344b8