Fix my profiler loader.
This is what happens when you test on a pre-configured environment that
already worked fine.    This depended on a function that was never called
so we ended up probing for paths on NULL (repeating the same process twice)
instead of loading the profiler from the installation directory.


Commit migrated from https://github.com/mono/mono/commit/8fdd8ebfe60b743938a02c6845c842203e456658