[sgen] Workers dequeue from the distribute gray queue directly.
Previously we manually fed the workers, only restarting the world
when the distribute gray queue was empty, which neglected most of the
benefit of concurrent collection.  Now we use locking for the queue and
let workers drain it while the world runs again.

If the locking turns out to be a bottleneck we could implement a lock-free
queue.


Commit migrated from https://github.com/mono/mono/commit/76087cedea5ab405ac34179c6933af8ede48f2d3