[sgen] Fix locking of the worker distribute gray queue.
The worker distribute gray queue is written to by the main GC thread
and read from by the worker thread(s) for work distribution.  Locking it is
therefore necessary, but we weren't doing it properly.

Previously I implemented optional locking for gray queues, but that's a
bad idea because it makes the enqueue fast path slower.  Now gray queues
are back to non-locking, but we have a new queue type, the section gray
queue, which is only accessible on a gray queue section granularity, not
on an individual object granularity, like regular gray queues.  Performance
there is not as critical, so locking is no problem.


Commit migrated from https://github.com/mono/mono/commit/3f011fdb1d5afcb204821af8628830e7c4cbd050