Include <stdint.h> to allow cross-compilation from OS X.
Cross-compiling mono on an OS X host for a Windows target using
mingw-w64 fails because of undefined __int64 types in a couple of places.

Including <stdint.h> fixes this and allows the OS X host to cross-compile
a Windows mono.

It probably sounds like an obscure use-case, but the wine-mono
project does just this, so ability to cross-compile for Windows on OS X
is crucial to being able to build wine-mono on a Mac.


Commit migrated from https://github.com/mono/mono/commit/9fa235bc34834ab7e37893c0661c7ae3a119cd43