[runtime]Provide a consistent view of registered threads during suspend operations.
We previously failed to provide a consistent view of the registered threads set. This
would should up in the following ways:

-New threads would join in between two FOREACH_THREAD passes. The recently joined threads
lead to broken invariants that one loop carried to the other.

-Existing threads would leave during a FOREACH_THREAD. The now defunct threads lead
to code waiting them to ack back and deadlock in the process.

-It was possible to witness an in-progress cleanup. This would show up by
mono_thread_info_current returning null when it should not.

To amend this we provide a a consistent view of all registered threads once the global
suspend lock is acquired. This is done by acquiring that lock when registering/unregistering.

This allows us to remove some of the work-around in the sgen thread registration
code and avoid taking the GC lock.


Commit migrated from https://github.com/mono/mono/commit/a1d20a2d53dbc15f5f334ba5889382ecd223b0bf