[runtime]Type parsing of generics should not rely on the presence of a backtick. Fixes mono/mono#15124.
We previously depended on the presence of a backtick in a generic type name to try to
parse a generics argument list. This is not what the spec says and F# breaks this expectation.

The right way to handle this is to probe for array syntax.


Commit migrated from https://github.com/mono/mono/commit/bd741f0aacfd7ce246aec35e19e1ffac462bb1dd