[sgen]Fix a deadlock when the GC try to free a JI during world restart. Fixes mono/mono#15759
This is deadlock prone as the GC lock is a leaf lock while the domain
lock required to free a JI is not.

There are plenty of places where we alloc memory while holding the
domain lock, so we cannot do it the other way.


Commit migrated from https://github.com/mono/mono/commit/971eba81ac3b150f806f5414e10ae4808ebf92ad