Clean up mono/tests/Makefile.am.
This removes a lot of duplicated logic and special-casing for Wrench. There
are now only 3 important top-level targets:

* `check`: Run all stable (~fast) tests.
* `check-full`: Run all stable, unstable, and slow tests.
* `check-parallel`: Compile tests in parallel, then run `check-full`.

By default, `no-dump` is passed to the test runner. To pass `dump-output` and
get `stdout`/`stderr` dumped to console when a failure occurs, set `V` as in:

    $ make check V=1

Or:

    $ make check-full V=1

If you're running tests in a CI environment, it may be desirable to disable
tests that are known to be unstable. To do so, run:

    $ make check CI=1

Or:

    $ make check-full CI=1


Commit migrated from https://github.com/mono/mono/commit/a7712d90931a29c76f83e6275246c288dee38644