Make assembly resolve events more compatible with the MS runtime.
We now try and trigger assembly resolve events for referenced assemblies
in a way that is more compatible with the MS runtime behaviour. If that still
fails we continue falling back and looking in the same directory as the referencing
image as before to keep compatibility with code that relies on the previous mono
behaviour.
The assemblyresolve_event3.cs test that used to fail on the MS runtime has been fixed
so it works now with both engines.


Commit migrated from https://github.com/mono/mono/commit/aa181455bb318523dbd0cf78d60275314bb24308