[MWF] Implement multi-display support on Linux (Fixes mono/mono#325669)
The number of displays and sizes can be retrieved with the help of
libXinerama. The implementation was done in a way that is backwards
compatible, so on non-Linux systems or when libXinerama is not
available the behavior will be the same as before.

Also fixed the implementation of FormStartPosition.CenterScreen to
bring up the form centered on the current display (with the current
display being either the screen the top left corner of the owner
form is on, or if there is no owner the screen that has the mouse
pointer). This fixes Novell bug mono/mono#325669 on Linux
(https://bugzilla.novell.com/show_bug.cgi?id=325669).


Commit migrated from https://github.com/mono/mono/commit/e1b7ecdb7b9d58b3aac97e739cd30d561ffa2ff1