[gc] Add finalizer API, implemented only on sgen.
The new finalizer API allows an embedder to be notified during a collector
when an object is queued for finalization.

This is a VERY low level interface, as this is called with the world stopped, so the
code MUST be async safe. If you don't precisely know what that means, don't use it.


Commit migrated from https://github.com/mono/mono/commit/a153eed8354a602b8fe02c8e5a9b9919d85067f7