[runtime] When parsing a method header, ensure we pass any available generic context. Fixes mono/mono#21388.
We need to pass the right generic context to the method parsing so local variables with
generic argument types get the owner properly set.

In addition to that, harden the icall in case the owner is not set.


Commit migrated from https://github.com/mono/mono/commit/19940950495764be30a85b91392dbaed8e05af35