[sgen] Mark on dequeue, not enqueue.
This is another experiment.

When marking on enqueue, what we need to do is look at the object
we're about to enqueue, to check whether it's a small or large object.
Then, when we dequeue to scan, we look at the object again, so we
essentially fetch each live object header twice.

With mark on dequeue, when we encounter a pointer we always enqueue
it, without even looking at the object.  Only when we dequeue do we
figure out whether it's marked yet, and thus whether we need to scan
it.


Commit migrated from https://github.com/mono/mono/commit/ab83e6316d9acfda22c592458f341018ae92bd22